Moving iree/base/atomics.h to iree/base/internal/.
