+++ bindings/tflite/CMakeLists.txt
@@ -35,6 +35,7 @@ iree_cc_library(
   DEPS
     iree::base::api
     iree::base::core_headers
+    iree::base::internal
     iree::base::tracing
     iree::hal::api
     iree::hal::drivers
