+++ iree/base/threading_darwin.c
@@ -23,7 +23,7 @@
 #include <pthread.h>
 #include <string.h>
 
-#include "iree/base/atomics.h"
+#include "iree/base/internal/atomics.h"
 #include "iree/base/threading.h"
 #include "iree/base/tracing.h"
