+++ iree/base/threading_win32.c
@@ -19,7 +19,7 @@
 
 #include <stdio.h>
 
-#include "iree/base/atomics.h"
+#include "iree/base/internal/atomics.h"
 #include "iree/base/threading.h"
 #include "iree/base/threading_impl.h"
 #include "iree/base/tracing.h"
