+++ iree/hal/BUILD
@@ -77,6 +77,7 @@ cc_library(
         "//iree/base:synchronization",
         "//iree/base:threading",
         "//iree/base:tracing",
+        "//iree/base/internal",
         "//iree/hal/vulkan/util:ref_ptr",
         "@com_google_absl//absl/container:inlined_vector",
         "@com_google_absl//absl/strings",
