+++ iree/hal/CMakeLists.txt
@@ -63,6 +63,7 @@ iree_cc_library(
     half::includes
     iree::base::api
     iree::base::core_headers
+    iree::base::internal
     iree::base::synchronization
     iree::base::threading
     iree::base::tracing
