+++ iree/hal/resource.h
@@ -19,8 +19,8 @@
 #include <stdint.h>
 
 #include "iree/base/api.h"
-#include "iree/base/atomics.h"
 #include "iree/base/debugging.h"
+#include "iree/base/internal/atomics.h"
 
 #ifdef __cplusplus
 extern "C" {
