+++ iree/task/affinity_set.h
@@ -15,7 +15,7 @@
 #ifndef IREE_TASK_AFFINITY_SET_H_
 #define IREE_TASK_AFFINITY_SET_H_
 
-#include "iree/base/atomics.h"
+#include "iree/base/internal/atomics.h"
 #include "iree/base/math.h"
 #include "iree/task/tuning.h"
