+++ iree/task/executor.h
@@ -16,7 +16,7 @@
 #define IREE_TASK_EXECUTOR_H_
 
 #include "iree/base/api.h"
-#include "iree/base/atomics.h"
+#include "iree/base/internal/atomics.h"
 #include "iree/base/wait_handle.h"
 #include "iree/task/scope.h"
 #include "iree/task/submission.h"
