+++ iree/task/scope.h
@@ -16,7 +16,7 @@
 #define IREE_TASK_SCOPE_H_
 
 #include "iree/base/api.h"
-#include "iree/base/atomics.h"
+#include "iree/base/internal/atomics.h"
 #include "iree/base/synchronization.h"
 #include "iree/base/tracing.h"
 #include "iree/task/task.h"
