+++ iree/vm/BUILD
@@ -91,6 +91,7 @@ cc_library(
         "//iree/base:api",
         "//iree/base:core_headers",
         "//iree/base:tracing",
+        "//iree/base/internal",
     ],
 )
