+++ iree/vm/module.h
@@ -19,7 +19,7 @@
 
 #include "iree/base/alignment.h"
 #include "iree/base/api.h"
-#include "iree/base/atomics.h"
+#include "iree/base/internal/atomics.h"
 
 #ifdef __cplusplus
 extern "C" {
